CRUD of Read done
